pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbsavr"
description = "An open source tool for easily creating database backups and storing them in S3-compatible object storage"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
dynamic = ["version"]
keywords = ["database", "backup", "postgresql", "mysql", "mongodb", "s3"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Database",
    "Topic :: System :: Archiving :: Backup",
]
dependencies = [
    "boto3>=1.20.0",
    "celery>=5.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/dbsavr"
"Bug Tracker" = "https://github.com/yourusername/dbsavr/issues"

[project.scripts]
dbsavr = "dbsavr.cli:cli"

[tool.setuptools]
packages = ["dbsavr"]

[tool.setuptools.dynamic]
version = {attr = "dbsavr.version.__version__"}